test: keep junction registrations behavior-owned
