+++ packages/device-syncd/test/junction-provider-diagnostics.test.ts
@@ -1436,6 +1436,131 @@ test("Junction routine, precise, and retained calendar writers share persisted s
   );
 });
 
+const usefulHistoricalSummaryRecordByResource = {
+  sleep: {
+    id: "sleep-1",
+    connectionId: "provider-garmin-1",
+    startAt: "2026-04-02T01:00:00.000Z",
+    endAt: "2026-04-02T08:00:00.000Z",
+  },
+  workouts: {
+    id: "workouts-1",
+    connectionId: "provider-garmin-1",
+    startAt: "2026-04-02T12:00:00.000Z",
+    durationMinutes: 45,
+  },
+  body: {
+    id: "body-1",
+    connectionId: "provider-garmin-1",
+    weightKg: 72,
+  },
+} satisfies Record<"sleep" | "workouts" | "body", Record<string, unknown>>;
+
+const usefulHistoricalSummaryCompletionCases = [
+  {
+    label: "activity floors",
+    resource: "activity",
+    record: {
+      id: "activity-floors-1",
+      connectionId: "provider-garmin-1",
+      floorsClimbed: 8,
+    },
+  },
+  {
+    label: "body lean mass",
+    resource: "body",
+    record: {
+      id: "body-lean-1",
+      connectionId: "provider-garmin-1",
+      leanBodyMassKg: 58.2,
+    },
+  },
+  {
+    label: "body waist circumference",
+    resource: "body",
+    record: {
+      id: "body-waist-1",
+      connectionId: "provider-garmin-1",
+      waistCircumferenceCm: 82,
+    },
+  },
+  {
+    label: "meal raw-only",
+    resource: "meal",
+    record: {
+      id: "meal-1",
+      sourceProviderSlug: "garmin",
+      mealType: "breakfast",
+    },
+  },
+  {
+    label: "menstrual cycle raw-only",
+    resource: "menstrual_cycle",
+    record: {
+      id: "cycle-1",
+      sourceProviderSlug: "garmin",
+      cycleDay: 3,
+    },
+  },
+] as const;
+
+for (const testCase of usefulHistoricalSummaryCompletionCases) {
+  test(`Junction ${testCase.label} summary historical backfill marks the historical window complete`, async () => {
+    const importedSnapshots: unknown[] = [];
+    const provider = createJunctionProvider(async (input) => {
+      const url = readUrl(input);
+
+      if (url === "https://api.sandbox.us.junction.com/v2/user/providers/junction-user-1") {
+        return createJsonResponse({
+          providers: [
+            {
+              id: "provider-garmin-1",
+              slug: "garmin",
+              name: "Garmin",
+              status: "connected",
+              resource_availability: {
+                [testCase.resource]: true,
+                heartrate: true,
+              },
+            },
+          ],
+        });
+      }
+
+      if (url.startsWith(`https://api.sandbox.us.junction.com/v2/summary/${testCase.resource}/junction-user-1`)) {
+        return createJsonResponse({ data: [testCase.record] });
+      }
+      throw new Error(`Unexpected request: ${url}`);
+    }, {
+      summaryResources: [testCase.resource],
+    });
+
+    const result = await executeJunctionJob(
+      provider,
+      createJunctionJobContext({
+        importSnapshot: async (snapshot) => {
+          importedSnapshots.push(snapshot);
+          return { imported: true };
+        },
+      }),
+      createJob("backfill", {
+        windowStart: "2026-04-01T00:00:00.000Z",
+        windowEnd: "2026-04-03T00:00:00.000Z",
+      }),
+    );
+
+    assert.deepEqual(result.metadataPatch, {
+      junctionHistoricalBackfillStatus: "coverage_v3_complete",
+      junctionHistoricalBackfillEmptyAttempts: 0,
+      junctionHistoricalBackfillLastEmptyAt: null,
+      junctionHistoricalBackfillWindowStart: "2026-04-01T00:00:00.000Z",
+      junctionHistoricalBackfillWindowEnd: "2026-04-03T00:00:00.000Z",
+    });
+    assert.equal(result.scheduledJobs, undefined);
+    assert.equal(importedSnapshots.length, 1);
+  });
+}
+
 test("Junction summary reads extract the documented top-level meals envelope", async () => {
   const importedSnapshots: unknown[] = [];
   const provider = createJunctionProvider(async (input) => {
@@ -1499,6 +1624,173 @@ test("Junction summary reads extract the documented top-level meals envelope", a
   assert.equal(mealRecord?.id, "meal-doc-envelope-1");
   assert.equal(Object.hasOwn(mealRecord ?? {}, "meals"), false);
 });
+
+for (const testCase of [
+  {
+    label: "meal identity/debug-only",
+    resource: "meal",
+    record: {
+      id: "meal-provenance-only",
+      clientId: "provider-client-1",
+      debug: true,
+      fullName: "Raw Member Name",
+      patientName: "Raw Patient Name",
+      sourceProviderSlug: "garmin",
+      status: "synced",
+      items: [
+        {
+          subjectId: "raw-meal-subject-id",
+          subject: {
+            id: "raw-meal-nested-subject-id",
+          },
+        },
+      ],
+    },
+  },
+  {
+    label: "menstrual cycle identity/contact-only",
+    resource: "menstrual_cycle",
+    record: {
+      addressLine1: "123 Private Street",
+      birthDate: "1980-01-01",
+      dateOfBirth: "1980-01-01",
+      dob: "1980-01-01",
+      id: "cycle-identity-only",
+      memberName: "Raw Member Name",
+      provider_connection_id: "provider-garmin-1",
+      sourceProviderSlug: "garmin",
+      user: {
+        id: "raw-cycle-user-id",
+      },
+      profile: {
+        patient_id: "raw-cycle-patient-id",
+      },
+      symptoms: [
+        {
+          subjectId: "raw-cycle-subject-id",
+          subjects: [
+            {
+              id: "raw-cycle-subjects-container-id",
+            },
+          ],
+        },
+      ],
+    },
+  },
+] as const) {
+  test(`Junction ${testCase.label} raw-only summary does not create a historical obligation`, async () => {
+    const importedSnapshots: unknown[] = [];
+    const provider = createJunctionProvider(async (input) => {
+      const url = readUrl(input);
+
+      if (url === "https://api.sandbox.us.junction.com/v2/user/providers/junction-user-1") {
+        return createJsonResponse({
+          providers: [
+            {
+              id: "provider-garmin-1",
+              slug: "garmin",
+              name: "Garmin",
+              status: "connected",
+              resource_availability: {
+                [testCase.resource]: true,
+                heartrate: true,
+              },
+            },
+          ],
+        });
+      }
+
+      if (url.startsWith(`https://api.sandbox.us.junction.com/v2/summary/${testCase.resource}/junction-user-1`)) {
+        return createJsonResponse({ data: [testCase.record] });
+      }
+      throw new Error(`Unexpected request: ${url}`);
+    }, {
+      summaryResources: [testCase.resource],
+    });
+
+    const result = await executeJunctionJob(
+      provider,
+      createJunctionJobContext({
+        now: "2026-04-04T00:00:00.000Z",
+        importSnapshot: async (snapshot) => {
+          importedSnapshots.push(snapshot);
+          return { imported: true };
+        },
+      }),
+      createJob("backfill", {
+        windowStart: "2026-04-01T00:00:00.000Z",
+        windowEnd: "2026-04-03T00:00:00.000Z",
+      }),
+    );
+
+    assert.deepEqual(result.metadataPatch, {
+      junctionHistoricalBackfillStatus: "coverage_v3_complete",
+      junctionHistoricalBackfillEmptyAttempts: 0,
+      junctionHistoricalBackfillLastEmptyAt: null,
+      junctionHistoricalBackfillWindowStart: "2026-04-01T00:00:00.000Z",
+      junctionHistoricalBackfillWindowEnd: "2026-04-03T00:00:00.000Z",
+    });
+    assert.equal(result.scheduledJobs, undefined);
+    assert.equal(importedSnapshots.length, 1);
+  });
+}
+
+for (const summaryResource of ["sleep", "workouts", "body"] as const) {
+  test(`Junction ${summaryResource} summary historical backfill marks the historical window complete`, async () => {
+    const importedSnapshots: unknown[] = [];
+    const provider = createJunctionProvider(async (input) => {
+      const url = readUrl(input);
+
+      if (url === "https://api.sandbox.us.junction.com/v2/user/providers/junction-user-1") {
+        return createJsonResponse({
+          providers: [
+            {
+              id: "provider-garmin-1",
+              slug: "garmin",
+              name: "Garmin",
+              status: "connected",
+              resource_availability: {
+                [summaryResource]: true,
+                heartrate: true,
+              },
+            },
+          ],
+        });
+      }
+
+      if (url.startsWith(`https://api.sandbox.us.junction.com/v2/summary/${summaryResource}/junction-user-1`)) {
+        return createJsonResponse({ data: [usefulHistoricalSummaryRecordByResource[summaryResource]] });
+      }
+      throw new Error(`Unexpected request: ${url}`);
+    }, {
+      summaryResources: [summaryResource],
+    });
+
+    const result = await executeJunctionJob(
+      provider,
+      createJunctionJobContext({
+        importSnapshot: async (snapshot) => {
+          importedSnapshots.push(snapshot);
+          return { imported: true };
+        },
+      }),
+      createJob("backfill", {
+        windowStart: "2026-04-01T00:00:00.000Z",
+        windowEnd: "2026-04-03T00:00:00.000Z",
+      }),
+    );
+
+    assert.deepEqual(result.metadataPatch, {
+      junctionHistoricalBackfillStatus: "coverage_v3_complete",
+      junctionHistoricalBackfillEmptyAttempts: 0,
+      junctionHistoricalBackfillLastEmptyAt: null,
+      junctionHistoricalBackfillWindowStart: "2026-04-01T00:00:00.000Z",
+      junctionHistoricalBackfillWindowEnd: "2026-04-03T00:00:00.000Z",
+    });
+    assert.equal(result.scheduledJobs, undefined);
+    assert.equal(importedSnapshots.length, 1);
+  });
+}
 
 test("Junction sleep_cycle stage-count-only history stays pending without canonical evidence", async () => {
   const importedSnapshots: unknown[] = [];
@@ -1567,6 +1859,131 @@ test("Junction sleep_cycle stage-count-only history stays pending without canoni
   );
   assertJunctionWindowQuery(summaryRequest, "2026-04-01", "2026-04-02");
 });
+
+for (const summaryResource of ["activity", "sleep"] as const) {
+  test(`Junction ${summaryResource} id-only historical backfill keeps the summary window retrying`, async () => {
+    const importedSnapshots: unknown[] = [];
+    const provider = createJunctionProvider(async (input) => {
+      const url = readUrl(input);
+
+      if (url === "https://api.sandbox.us.junction.com/v2/user/providers/junction-user-1") {
+        return createJsonResponse({
+          providers: [
+            {
+              id: "provider-garmin-1",
+              slug: "garmin",
+              name: "Garmin",
+              status: "connected",
+              resource_availability: {
+                [summaryResource]: true,
+                heartrate: true,
+              },
+            },
+          ],
+        });
+      }
+
+      if (url.startsWith(`https://api.sandbox.us.junction.com/v2/summary/${summaryResource}/junction-user-1`)) {
+        return createJsonResponse({ data: [{ id: `${summaryResource}-1`, connectionId: "provider-garmin-1" }] });
+      }
+      throw new Error(`Unexpected request: ${url}`);
+    }, {
+      summaryResources: [summaryResource],
+    });
+
+    const result = await executeJunctionJob(
+      provider,
+      createJunctionJobContext({
+        now: "2026-04-04T00:00:00.000Z",
+        importSnapshot: async (snapshot) => {
+          importedSnapshots.push(snapshot);
+          return { imported: true };
+        },
+      }),
+      createJob("backfill", {
+        windowStart: "2026-04-01T00:00:00.000Z",
+        windowEnd: "2026-04-03T00:00:00.000Z",
+      }),
+    );
+
+    assert.deepEqual(result.metadataPatch, {
+      junctionHistoricalBackfillStatus: "coverage_v3_retrying",
+      junctionHistoricalBackfillEmptyAttempts: 1,
+      junctionHistoricalBackfillLastEmptyAt: "2026-04-04T00:00:00.000Z",
+      junctionHistoricalBackfillWindowStart: "2026-04-01T00:00:00.000Z",
+      junctionHistoricalBackfillWindowEnd: "2026-04-03T00:00:00.000Z",
+    });
+    assertConnectBackfillRetryWake(result, "2026-04-04T00:15:00.000Z");
+    assert.equal(importedSnapshots.length, 1);
+  });
+}
+
+const floatingSessionOnlySummaryRecordByResource = {
+  sleep: {
+    id: "sleep-1",
+    connectionId: "provider-garmin-1",
+    startAt: "2026-04-02T01:00:00",
+    endAt: "2026-04-02T08:00:00",
+  },
+} satisfies Record<"sleep", Record<string, unknown>>;
+
+for (const summaryResource of ["sleep"] as const) {
+  test(`Junction ${summaryResource} floating session-only historical backfill keeps the summary window retrying`, async () => {
+    const importedSnapshots: unknown[] = [];
+    const provider = createJunctionProvider(async (input) => {
+      const url = readUrl(input);
+
+      if (url === "https://api.sandbox.us.junction.com/v2/user/providers/junction-user-1") {
+        return createJsonResponse({
+          providers: [
+            {
+              id: "provider-garmin-1",
+              slug: "garmin",
+              name: "Garmin",
+              status: "connected",
+              resource_availability: {
+                [summaryResource]: true,
+                heartrate: true,
+              },
+            },
+          ],
+        });
+      }
+
+      if (url.startsWith(`https://api.sandbox.us.junction.com/v2/summary/${summaryResource}/junction-user-1`)) {
+        return createJsonResponse({ data: [floatingSessionOnlySummaryRecordByResource[summaryResource]] });
+      }
+      throw new Error(`Unexpected request: ${url}`);
+    }, {
+      summaryResources: [summaryResource],
+    });
+
+    const result = await executeJunctionJob(
+      provider,
+      createJunctionJobContext({
+        now: "2026-04-04T00:00:00.000Z",
+        importSnapshot: async (snapshot) => {
+          importedSnapshots.push(snapshot);
+          return { imported: true };
+        },
+      }),
+      createJob("backfill", {
+        windowStart: "2026-04-01T00:00:00.000Z",
+        windowEnd: "2026-04-03T00:00:00.000Z",
+      }),
+    );
+
+    assert.deepEqual(result.metadataPatch, {
+      junctionHistoricalBackfillStatus: "coverage_v3_retrying",
+      junctionHistoricalBackfillEmptyAttempts: 1,
+      junctionHistoricalBackfillLastEmptyAt: "2026-04-04T00:00:00.000Z",
+      junctionHistoricalBackfillWindowStart: "2026-04-01T00:00:00.000Z",
+      junctionHistoricalBackfillWindowEnd: "2026-04-03T00:00:00.000Z",
+    });
+    assertConnectBackfillRetryWake(result, "2026-04-04T00:15:00.000Z");
+    assert.equal(importedSnapshots.length, 1);
+  });
+}
 
 test("Junction useful summary without source linkage keeps the historical window retrying", async () => {
   const importedSnapshots: unknown[] = [];
